feat(infrastructure): extract infrastructure module into modules/infrastructure/ (#708)
Move 7 services (edge-capability-guard, edge-async-log-fetcher, edge-log-fetcher,
docker-frame-decoder, elasticsearch-config, elasticsearch-log-forwarder, kibana-client)
and 1 route (edge-jobs) into backend/src/modules/infrastructure/.

Create barrel export (services-only, routes excluded to avoid eager-loading TDZ issues)
and route aggregator. Update all 7 consumer files to import from the new module.
