+++ backend/src/modules/infrastructure/routes/index.ts
@@ -1,0 +1,6 @@
+import { FastifyInstance } from 'fastify';
+import { edgeJobsRoutes } from './edge-jobs.js';
+
+export async function infrastructureRoutes(fastify: FastifyInstance) {
+  await fastify.register(edgeJobsRoutes);
+}
