+++ backend/src/modules/infrastructure/services/docker-frame-decoder.ts
