+++ backend/src/routes/logs.ts
@@ -2,7 +2,7 @@ import { FastifyInstance } from 'fastify';
 import { Agent } from 'undici';
 import { createChildLogger } from '../core/utils/logger.js';
 import { LogsSearchQuerySchema, LogsTestBodySchema } from '../core/models/api-schemas.js';
-import { getElasticsearchConfig } from '../services/elasticsearch-config.js';
+import { getElasticsearchConfig } from '../modules/infrastructure/index.js';
 
 const log = createChildLogger('logs-route');
